requirements.txt
# LiveKit bridge
livekit>=1.0.0
python-dotenv>=1.0.0

# LiveKit voice agent (only needed for livekit_voice_agent.py)
livekit-agents>=1.4
livekit-plugins-openai>=1.0
livekit-plugins-silero>=1.0

numpy==2.2.6
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
scipy==1.16.3
matplotlib==3.10.7
gamepad==0.1.1
PyOpenGL==3.1.10
glfw==2.10.0

omegaconf==2.3.0
PyYAML==6.0.3
joblib==1.5.2
easydict==1.13
packaging==25.0

tqdm==4.67.1
typing_extensions==4.15.0

transformers==4.56.2
accelerate==1.10.1
deepspeed==0.18.0
peft==0.17.1
safetensors==0.6.2
tokenizers==0.22.1
sentencepiece==0.2.1
einops==0.8.1

huggingface-hub==0.35.1
datasets==4.2.0
modelscope==1.30.0
hf-xet==1.1.10

nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
triton==3.4.0

aiohttp==3.13.0
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
requests==2.32.5
urllib3==2.5.0
yarl==1.22.0
multidict==6.7.0
frozenlist==1.8.0

msgpack==1.1.2
protobuf==6.32.1
pyarrow==21.0.0
pandas==2.3.3
hjson==3.1.0

opencv-python-headless==4.12.0.88
pillow==11.0.0
av==15.1.0
decord==0.6.0

psutil==7.1.0
py-cpuinfo==9.0.0
multiprocess==0.70.16
dill==0.4.0

pytest==8.4.2
pytest-asyncio==1.2.0
absl-py==2.3.1

colorlog==6.9.0
wandb==0.22.2
sentry-sdk==2.41.0
click==8.3.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
six==1.17.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
sniffio==1.3.1
anyio==4.11.0
regex==2025.9.18
xxhash==3.6.0
zipp==3.23.0

pydantic==2.12.0
pydantic_core==2.41.1
annotated-types==0.7.0
typing-inspection==0.4.2

sympy==1.13.3
mpmath==1.3.0
networkx==3.3

filelock==3.19.1
importlib_resources==6.5.2
platformdirs==4.5.0

GitPython==3.1.45
gitdb==4.0.12
smmap==5.0.2

redis==6.4.0

etils==1.13.0
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
kiwisolver==1.4.9
pyparsing==3.2.5
Pygments==2.19.2
propcache==0.4.1
iniconfig==2.1.0
pluggy==1.6.0
setuptools==78.1.1
wheel==0.45.1

qwen-vl-utils==0.0.14
antlr4-python3-runtime==4.9.3
ninja==1.13.0
